refactor: extract last state of gametable to component EndGame
