+++ src/components/GameTable/components/EndGame/style.css
